+++ src/components/progress-bar/progress-bar.ts
@@ -43,7 +43,8 @@ export default class SlProgressBar extends ShoelaceElement {
         part="base"
         class=${classMap({
           'progress-bar': true,
-          'progress-bar--indeterminate': this.indeterminate
+          'progress-bar--indeterminate': this.indeterminate,
+          'progress-bar--rtl': this.localize.dir() === 'rtl'
         })}
         role="progressbar"
         title=${ifDefined(this.title)}